define columns in ParticipantListSection component
