+++ src/features/events/components/EventParticipantsList.tsx
@@ -1,17 +1,13 @@
+import { Box } from '@mui/material';
 import { FC } from 'react';
-import { GridColDef } from '@mui/x-data-grid-pro';
-import { Box, Button, Typography } from '@mui/material';
 
 import AddPersonButton from './AddPersonButton';
 import EventDataModel from 'features/events/models/EventDataModel';
 import messageIds from 'features/events/l10n/messageIds';
-import noPropagate from 'utils/noPropagate';
 import ParticipantListSection from 'features/events/components/ParticipantListSection';
 import theme from 'theme';
 import { useMessages } from 'core/i18n';
 import { ZetkinEvent } from 'utils/types/zetkin';
-import ZUIAvatar from 'zui/ZUIAvatar';
-import ZUIRelativeTime from 'zui/ZUIRelativeTime';
 
 interface EventParticipantsListProps {
   data: ZetkinEvent;
@@ -26,125 +22,6 @@ const EventParticipantsList: FC<EventParticipantsListProps> = ({
 }) => {
   const messages = useMessages(messageIds);
 
-  const columns: GridColDef[] = [
-    {
-      align: 'center',
-      disableColumnMenu: true,
-      field: 'avatar',
-      headerName: '',
-      hideSortIcons: true,
-      renderCell: (params) => (
-        <ZUIAvatar orgId={orgId} personId={params.row.id} />
-      ),
-      resizable: false,
-      sortable: false,
-      width: 20,
-    },
-    {
-      disableColumnMenu: true,
-      field: 'name',
-      flex: 1,
-      headerName: messages.eventParticipantsList.columnName(),
-      hideSortIcons: true,
-      renderCell: (params) => {
-        if (params.row.person) {
-          return <Typography>{params.row.person.name}</Typography>;
-        } else {
-          return (
-            <Typography>
-              {params.row.first_name + ' ' + params.row.last_name}
-            </Typography>
-          );
-        }
-      },
-      resizable: false,
-      sortable: false,
-    },
-    {
-      disableColumnMenu: true,
-      field: 'phone',
-      flex: 1,
-      headerName: messages.eventParticipantsList.columnPhone(),
-      hideSortIcons: true,
-      renderCell: (params) => {
-        if (params.row.person) {
-          return <Typography>{''}</Typography>;
-        } else {
-          return <Typography>{params.row.phone}</Typography>;
-        }
-      },
-      resizable: false,
-      sortable: false,
-    },
-    {
-      disableColumnMenu: true,
-      field: 'email',
-      flex: 1,
-      headerName: messages.eventParticipantsList.columnEmail(),
-      hideSortIcons: true,
-      resizable: false,
-      sortable: false,
-      valueGetter: (params) => {
-        if (params.row.person) {
-          return '';
-        } else {
-          return `${params.row.email}`;
-        }
-      },
-    },
-    {
-      disableColumnMenu: true,
-      field: 'notified',
-      flex: 1,
-      headerName: messages.eventParticipantsList.columnNotified(),
-      hideSortIcons: true,
-      renderCell: (params) => {
-        if (params.row.person) {
-          return <ZUIRelativeTime datetime={params.row.response_date} />;
-        } else {
-          return <ZUIRelativeTime datetime={params.row.reminder_sent} />;
-        }
-      },
-      resizable: false,
-      sortable: false,
-    },
-    {
-      align: 'right',
-      disableColumnMenu: true,
-      field: 'cancel',
-      flex: 1,
-      headerName: '',
-      hideSortIcons: true,
-      renderCell: (params) => {
-        if (params.row.person) {
-          return (
-            <>
-              <Button sx={{ marginRight: '10px' }} variant="text">
-                {messages.eventParticipantsList.buttonCancel()}
-              </Button>
-              <Button
-                onClick={noPropagate(() => {
-                  model.addParticipant(params.row.id);
-                })}
-                variant="outlined"
-              >
-                {messages.eventParticipantsList.buttonBook()}
-              </Button>
-            </>
-          );
-        } else {
-          return (
-            <Button variant="text">
-              {messages.eventParticipantsList.buttonCancel()}
-            </Button>
-          );
-        }
-      },
-      resizable: false,
-      sortable: false,
-    },
-  ];
-
   return (
     <>
       <Box sx={{ display: 'flex', justifyContent: 'end' }}>
@@ -155,8 +32,9 @@ const EventParticipantsList: FC<EventParticipantsListProps> = ({
           <ParticipantListSection
             chipColor={theme.palette.grey[500]}
             chipNumber={model.getNumSignedParticipants().toString()}
-            columns={columns}
             description={messages.eventParticipantsList.descriptionSignups()}
+            model={model}
+            orgId={orgId}
             rows={model.getPendingSignUps() ?? []}
             title={messages.eventParticipantsList.signUps()}
           />
@@ -168,8 +46,9 @@ const EventParticipantsList: FC<EventParticipantsListProps> = ({
             '/' +
             data.num_participants_required
           }
-          columns={columns}
           description={messages.eventParticipantsList.descriptionBooked()}
+          model={model}
+          orgId={orgId}
           rows={model.getParticipants().data ?? []}
           title={messages.eventParticipantsList.bookedParticipants()}
         />
